set up pug, email service, added test template
