+++ consts/emailSubject.js
@@ -1,0 +1,5 @@
+const emailSubject = {
+  EMAIL_CONFIRMATION: 'EMAIL_CONFIRMATION'
+}
+
+module.exports = emailSubject
