+++ emails/confirm-email.pug
@@ -1,0 +1,6 @@
+extends index
+block headerText
+    h1 Welcome to Space2Study
+block content
+        p Thanks for signing up. To activate your account, please verify your email address by clicking the button below:
+            a(href="http://localhost:3000/activate/" + activationLink) activate your account
