+++ emails/index.pug
@@ -1,0 +1,12 @@
+doctype html
+html
+    head
+        meta(charset='UTF-8')
+        meta(name='viewport' content='width=device-width, initial-scale=1')
+    body
+        header
+            block headerText
+        main
+            block content
+        footer
+            p Best wishes,#[br] Space2Study:)
