+++ package.json
@@ -31,6 +31,7 @@
     "cookie-parser": "^1.4.6",
     "cors": "^2.8.5",
     "dotenv": "^10.0.0",
+    "email-templates": "^10.0.0",
     "express": "^4.17.1",
     "googleapis": "^105.0.0",
     "jsonwebtoken": "^8.5.1",
@@ -38,6 +39,7 @@
     "mongodb": "^4.2.2",
     "mongoose": "^6.1.2",
     "nodemailer": "^6.7.7",
+    "pug": "^3.0.2",
     "swagger-jsdoc": "^6.1.0",
     "swagger-ui-express": "^4.3.0",
     "uuid": "^8.3.2",
