+++ services/auth.js
@@ -11,6 +11,8 @@ const {
   USER_NOT_REGISTERED,
   PASSWORD_LENGTH_VALIDATION_FAILED
 } = require('~/consts/errors')
+const emailSubject = require('~/consts/emailSubject')
+const { sendMail } = require('~/utils/emailService')
 
 const authService = {
   signup: async (role, firstName, lastName, email, password) => {
@@ -31,6 +33,7 @@ const authService = {
 
     //TODO
     //await mailService.sendActivationMail(email, `${process.env.SERVER_URL}/api/activate/${activationLink}`)
+    await sendMail(email, emailSubject.EMAIL_CONFIRMATION, { activationLink })
 
     return {
       userEmail: user.email
